Remove the px4-v2_upstream configuration.  Moved to misc/Obsoleted/configs.
This is not the official configuration for the PX4 board and has led to confusion by NuttX users.  The board configuration also requires some ongoing maintenance and customization to support ongoing PX4 testing and evaluation.  It is best retained the PX4 repositories where it can be properly maintained and not in the upstream NuttX repository.